# docs/swagger.yaml
basePath: /
consumes:
- application/json
definitions:
  dto.ProjectDTO:
    properties:
      description:
        type: string
      done:
        type: boolean
      title:
        type: string
    required:
    - title
    type: object
  dto.SignInDTO:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  dto.SignUpDTO:
    properties:
      email:
        type: string
      name:
        minLength: 2
        type: string
      password:
        minLength: 8
        type: string
      username:
        minLength: 3
        type: string
    required:
    - email
    - name
    - password
    - username
    type: object
  handlers.errResponse:
    properties:
      message:
        type: string
    type: object
  handlers.statusResponse:
    properties:
      message:
        type: string
    type: object
host: localhost:8000
info:
  contact: {}
  title: Documentation for api
  version: "1.0"
paths:
  /api/projects:
    get:
      consumes:
      - application/json
      description: get project by id
      parameters:
      - description: project id
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ProjectDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.errResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.errResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handlers.errResponse'
      security:
      - ApiKeyAuth: []
      summary: GetById
      tags:
      - projects
  /api/projects/:
    get:
      consumes:
      - application/json
      description: get all projects
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.ProjectDTO'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.errResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handlers.errResponse'
      security:
      - ApiKeyAuth: []
      summary: GetAll
      tags:
      - projects
    post:
      consumes:
      - application/json
      description: create new project
      parameters:
      - description: project info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.ProjectDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.errResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.errResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handlers.errResponse'
      security:
      - ApiKeyAuth: []
      summary: Create
      tags:
      - projects
  /api/projects/{id}:
    delete:
      consumes:
      - application/json
      description: delete project by id
      parameters:
      - description: project id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.statusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.errResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.errResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handlers.errResponse'
      security:
      - ApiKeyAuth: []
      summary: DeleteById
      tags:
      - projects
    post:
      consumes:
      - application/json
      description: update project by id
      parameters:
      - description: project id
        in: path
        name: id
        required: true
        type: integer
      - description: project info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.ProjectDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.statusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.errResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.errResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handlers.errResponse'
      security:
      - ApiKeyAuth: []
      summary: UpdateById
      tags:
      - projects
  /auth/refresh:
    get:
      consumes:
      - application/json
      description: refreshing jwt
      parameters:
      - description: refresh token from cookie
        in: header
        name: refresh-token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.errResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handlers.errResponse'
      summary: refresh
      tags:
      - auth
  /auth/sign-in:
    post:
      consumes:
      - application/json
      parameters:
      - description: user details
        in: body
        name: SignInDTO
        required: true
        schema:
          $ref: '#/definitions/dto.SignInDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.errResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.errResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handlers.errResponse'
      summary: signIn
      tags:
      - auth
  /auth/sign-up:
    post:
      consumes:
      - application/json
      parameters:
      - description: user details
        in: body
        name: SignUpDTO
        required: true
        schema:
          $ref: '#/definitions/dto.SignUpDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.errResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.errResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handlers.errResponse'
      summary: signUp
      tags:
      - auth
produces:
- application/json
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"